+++ packages/picmo/src/createPicker.ts
@@ -1,4 +1,3 @@
-import styleInject from 'style-inject';
 import { AppEvents } from './AppEvents';
 import { EmojiPicker } from './views/EmojiPicker';
 import { PickerOptions, CustomEmoji, EmojiRecord } from './types';
@@ -8,10 +7,9 @@ import { DataStore } from './data/DataStore';
 import { initDatabase } from './data/emojiData';
 import { Bundle } from './i18n/bundle';
 import { getOptions } from './options';
+import { createStyleInjector } from './injectStyles';
 
-import globalConfig from './globalConfig';
 import css from './styles/index.css?inline';
-let cssLoaded = false;
 
 function initData(options: PickerOptions): Promise<DataStore> {
   return initDatabase(options.locale, options.dataStore, options.messages, options.emojiData);
@@ -23,16 +21,15 @@ function getPickerId() {
   return `picmo-${Date.now()}-${pickerIndex++}`;
 }
 
+const styleInject = createStyleInjector();
+
 /**
  * Creates a new emoji picker.
  * @param options The options for the emoji picker.
  * @returns a Promise that resolves to the picker when it is ready.
  */
 export function createPicker(options: Partial<PickerOptions>): EmojiPicker {
-  if (globalConfig.injectStyles && !cssLoaded) {
-    styleInject(css);  
-    cssLoaded = true;
-  }
+  styleInject(css);
   
   const finalOptions = getOptions(options);
   const customEmojis: EmojiRecord[] = (finalOptions?.custom || []).map((custom: CustomEmoji) => ({