+++ packages/popup-picker/src/index.ts
@@ -1,5 +1,4 @@
-import styleInject from 'style-inject';
-import { PickerOptions, globalConfig } from 'picmo';
+import { PickerOptions, createStyleInjector } from 'picmo';
 import { PopupOptions } from './types';
 import { PopupPickerController } from './popupPicker';
 
@@ -8,13 +7,10 @@ import css from './styles/index.css?inline';
 export * from './types';
 export { PopupPickerController };
 
-let cssLoaded = false;
+const styleInject = createStyleInjector();
 
 export function createPopup(pickerOptions: Partial<PickerOptions>, popupOptions: Partial<PopupOptions>) {
-  if (globalConfig.injectStyles && !cssLoaded) {
-    styleInject(css);
-    cssLoaded = true;
-  }
+  styleInject(css);
 
   const popup = new PopupPickerController({
     autoFocus: 'auto',