+++ packages/renderer-twemoji/src/declarations.ts
@@ -1,9 +1,0 @@
-declare module '*.svg' {
-  const content: string;
-  export default content;
-}
-
-declare module '*.css?inline' {
-  const content: string;
-  export default content;
-}