+++ packages/renderer-twemoji/vite.config.js
@@ -1,27 +1,3 @@
-import { resolve } from 'path';
+import createConfig from '../../vite.common';
 
-import nesting from 'postcss-nesting';
-import mixins from 'postcss-mixins';
-
-export default {
-  css: {
-    postcss: {
-      plugins: [
-        mixins(),
-        nesting(),
-      ]
-    }
-  },
-  build: {
-    lib: {
-      entry: resolve(__dirname, 'src/index.ts'),
-      name: 'PicMo-twemoji',
-      fileName: 'index'
-    }
-  },
-  // server: {
-    // headers: {
-      // 'Content-Security-Policy': `default-src 'self'; connect-src *; style-src 'self'`
-    // }
-  // }
-}
+export default createConfig(__dirname, 'picmoTwemoji');